extract monitor related transformers into separate file
