+++ src/jvm/main/org/jetbrains/kotlinx/lincheck/transformation/LincheckClassVisitor.kt
@@ -113,7 +113,7 @@ internal class LincheckClassVisitor(
             mv = SynchronizedMethodTransformer(fileName, className, methodName, createAdapter(mv), classVersion)
         }
         mv = MethodCallTransformer(fileName, className, methodName, createAdapter(mv))
-        mv = MonitorEnterAndExitTransformer(fileName, className, methodName, createAdapter(mv))
+        mv = MonitorTransformer(fileName, className, methodName, createAdapter(mv))
         mv = WaitNotifyTransformer(fileName, className, methodName, createAdapter(mv))
         mv = ParkUnparkTransformer(fileName, className, methodName, createAdapter(mv))
         mv = ObjectCreationTrackerTransformer(fileName, className, methodName, createAdapter(mv))
@@ -152,136 +152,6 @@ internal class LincheckClassVisitor(
                 invokeStatic(Injections::storeCancellableContinuation)
             }
             super.visitMethodInsn(opcodeAndSource, className, methodName, descriptor, isInterface)
-        }
-    }
-
-    /**
-     * Adds invocations of ManagedStrategy methods before monitorenter and monitorexit instructions
-     */
-    private inner class MonitorEnterAndExitTransformer(
-        fileName: String,
-        className: String,
-        methodName: String,
-        adapter: GeneratorAdapter,
-    ) : ManagedStrategyMethodVisitor(fileName, className, methodName, adapter) {
-        override fun visitInsn(opcode: Int) = adapter.run {
-            when (opcode) {
-                MONITORENTER -> {
-                    invokeIfInTestingCode(
-                        original = { monitorEnter() },
-                        code = {
-                            loadNewCodeLocationId()
-                            invokeStatic(Injections::beforeLock)
-                            invokeBeforeEventIfPluginEnabled("lock")
-                            invokeStatic(Injections::lock)
-                        }
-                    )
-                }
-
-                MONITOREXIT -> {
-                    invokeIfInTestingCode(
-                        original = { monitorExit() },
-                        code = {
-                            loadNewCodeLocationId()
-                            invokeStatic(Injections::unlock)
-                            invokeBeforeEventIfPluginEnabled("unlock")
-                        }
-                    )
-                }
-
-                else -> visitInsn(opcode)
-            }
-        }
-    }
-
-    /**
-     * Replace "method(...) {...}" with "method(...) {synchronized(this) {...} }"
-     */
-    private inner class SynchronizedMethodTransformer(
-        fileName: String,
-        className: String,
-        methodName: String,
-        adapter: GeneratorAdapter,
-        private val classVersion: Int
-    ) : ManagedStrategyMethodVisitor(fileName, className, methodName, adapter) {
-        private val isStatic: Boolean = this.adapter.access and ACC_STATIC != 0
-        private val tryLabel = Label()
-        private val catchLabel = Label()
-
-        override fun visitCode() = adapter.run {
-            invokeIfInTestingCode(
-                original = {},
-                code = {
-                    loadSynchronizedMethodMonitorOwner()
-                    monitorExit()
-                }
-            )
-            visitLabel(tryLabel)
-            invokeIfInTestingCode(
-                original = {},
-                code = {
-                    loadSynchronizedMethodMonitorOwner()
-                    loadNewCodeLocationId()
-                    invokeStatic(Injections::beforeLock)
-                    invokeBeforeEventIfPluginEnabled("lock")
-                    invokeStatic(Injections::lock)
-                }
-            )
-            visitCode()
-        }
-
-        override fun visitMaxs(maxStack: Int, maxLocals: Int) = adapter.run {
-            visitLabel(catchLabel)
-            invokeIfInTestingCode(
-                original = {},
-                code = {
-                    loadSynchronizedMethodMonitorOwner()
-                    loadNewCodeLocationId()
-                    invokeStatic(Injections::unlock)
-                    invokeBeforeEventIfPluginEnabled("unlock")
-                    loadSynchronizedMethodMonitorOwner()
-                    monitorEnter()
-                }
-            )
-            throwException()
-            visitTryCatchBlock(tryLabel, catchLabel, catchLabel, null)
-            visitMaxs(maxStack, maxLocals)
-        }
-
-        override fun visitInsn(opcode: Int) = adapter.run {
-            when (opcode) {
-                ARETURN, DRETURN, FRETURN, IRETURN, LRETURN, RETURN -> {
-                    invokeIfInTestingCode(
-                        original = {},
-                        code = {
-                            loadSynchronizedMethodMonitorOwner()
-                            loadNewCodeLocationId()
-                            invokeStatic(Injections::unlock)
-                            invokeBeforeEventIfPluginEnabled("unlock")
-                            loadSynchronizedMethodMonitorOwner()
-                            monitorEnter()
-                        }
-                    )
-                }
-            }
-            visitInsn(opcode)
-        }
-
-        private fun loadSynchronizedMethodMonitorOwner() = adapter.run {
-            if (isStatic) {
-                val classType = getType("L$className;")
-                if (classVersion >= V1_5) {
-                    visitLdcInsn(classType)
-                } else {
-                    visitLdcInsn(classType.className)
-                    invokeInIgnoredSection {
-                        invokeStatic(CLASS_TYPE, CLASS_FOR_NAME_METHOD)
-
-                    }
-                }
-            } else {
-                loadThis()
-            }
         }
     }
 
@@ -602,107 +472,6 @@ internal class LincheckClassVisitor(
                 }
             )
         }
-    }
-
-    /**
-     * Adds invocations of ManagedStrategy methods before wait and after notify calls
-     */
-    private inner class WaitNotifyTransformer(
-        fileName: String,
-        className: String,
-        methodName: String,
-        adapter: GeneratorAdapter,
-    ) : ManagedStrategyMethodVisitor(fileName, className, methodName, adapter) {
-        override fun visitMethodInsn(opcode: Int, owner: String, name: String, desc: String, itf: Boolean) =
-            adapter.run {
-                if (opcode == INVOKEVIRTUAL) {
-                    when {
-                        isWait0(name, desc) -> {
-                            invokeIfInTestingCode(
-                                original = {
-                                    visitMethodInsn(opcode, owner, name, desc, itf)
-                                },
-                                code = {
-                                    loadNewCodeLocationId()
-                                    invokeStatic(Injections::beforeWait)
-                                    invokeBeforeEventIfPluginEnabled("wait")
-                                    invokeStatic(Injections::wait)
-                                }
-                            )
-                        }
-
-                        isWait1(name, desc) -> {
-                            invokeIfInTestingCode(
-                                original = {
-                                    visitMethodInsn(opcode, owner, name, desc, itf)
-                                },
-                                code = {
-                                    pop2() // timeMillis
-                                    loadNewCodeLocationId()
-                                    invokeStatic(Injections::beforeWait)
-                                    invokeBeforeEventIfPluginEnabled("wait 1")
-                                    invokeStatic(Injections::waitWithTimeout)
-                                }
-                            )
-                        }
-
-                        isWait2(name, desc) -> {
-                            invokeIfInTestingCode(
-                                original = {
-                                    visitMethodInsn(opcode, owner, name, desc, itf)
-                                },
-                                code = {
-                                    pop() // timeNanos
-                                    pop2() // timeMillis
-                                    loadNewCodeLocationId()
-                                    invokeStatic(Injections::beforeWait)
-                                    invokeBeforeEventIfPluginEnabled("wait 2")
-                                    invokeStatic(Injections::waitWithTimeout)
-                                }
-                            )
-                        }
-
-                        isNotify(name, desc) -> {
-                            invokeIfInTestingCode(
-                                original = {
-                                    visitMethodInsn(opcode, owner, name, desc, itf)
-                                },
-                                code = {
-                                    loadNewCodeLocationId()
-                                    invokeStatic(Injections::notify)
-                                    invokeBeforeEventIfPluginEnabled("notify")
-                                }
-                            )
-                        }
-
-                        isNotifyAll(name, desc) -> {
-                            invokeIfInTestingCode(
-                                original = {
-                                    visitMethodInsn(opcode, owner, name, desc, itf)
-                                },
-                                code = {
-                                    loadNewCodeLocationId()
-                                    invokeStatic(Injections::notifyAll)
-                                    invokeBeforeEventIfPluginEnabled("notifyAll")
-                                }
-                            )
-                        }
-
-                        else -> {
-                            visitMethodInsn(opcode, owner, name, desc, itf)
-                        }
-                    }
-                } else {
-                    visitMethodInsn(opcode, owner, name, desc, itf)
-                }
-            }
-
-        private fun isWait0(mname: String, desc: String) = mname == "wait" && desc == "()V"
-        private fun isWait1(mname: String, desc: String) = mname == "wait" && desc == "(J)V"
-        private fun isWait2(mname: String, desc: String) = mname == "wait" && desc == "(JI)V"
-
-        private fun isNotify(mname: String, desc: String) = mname == "notify" && desc == "()V"
-        private fun isNotifyAll(mname: String, desc: String) = mname == "notifyAll" && desc == "()V"
     }
 
 }
